refactor: rename types and composable files
