+++ frontend/app/pages/event/[eventId]/book/seats/(types)/event-payment.ts
@@ -1,3 +1,4 @@
+// TODO: refactor to features/
 import type { SelectedSeat } from './ticket'
 
 export interface CartItem {
