+++ frontend/app/pages/event/[eventId]/book/seats/(types)/ticket.ts
